+++ timeline/src/main/java/org/mpilone/vaadin/timeline/Animation.java
@@ -1,0 +1,46 @@
+package org.mpilone.vaadin.timeline;
+
+/**
+ *
+ * @author mpilone
+ */
+public class Animation {
+
+  private final Integer duration;
+  private final EasingFunction easingFunction;
+
+  public Animation() {
+    this(null, null);
+  }
+
+  public Animation(Integer duration, EasingFunction easingFunction) {
+    this.duration = duration;
+    this.easingFunction = easingFunction;
+  }
+
+  public Integer getDuration() {
+    return duration;
+  }
+
+  public EasingFunction getEasingFunction() {
+    return easingFunction;
+  }
+
+  public static enum EasingFunction {
+
+    linear,
+    easeInQuad,
+    easeOutQuad,
+    easeInOutQuad,
+    easeInCubic,
+    easeOutCubic,
+    easeInOutCubic,
+    easeInQuart,
+    easeOutQuart,
+    easeInOutQuart,
+    easeInQuint,
+    easeOutQuint,
+    easeInOutQuint;
+  }
+
+}